Attempting to develop a better square error rejection test.
It's a dilemma: Want to refine the neighborhood before we do
least squares fitting, but also want to keep the outliers to
determine the magnitude of the error.

My idea to improve this is to search for spatial locality in
not just the normal direction, but the tangential direction as
well. This could be done by a prepass of sorting the features
supporting a peak by distance, and removing duplicates. There
aren't that many features, so the performance hit from sorting
should be dominated by the sobel filter.